+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/TranslatorPass.php
@@ -11,9 +11,11 @@
 
 namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
 
+use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ServiceLocator;
 
 class TranslatorPass implements CompilerPassInterface
 {
@@ -24,7 +26,9 @@ class TranslatorPass implements CompilerPassInterface
         }
 
         $loaders = array();
+        $loaderRefs = array();
         foreach ($container->findTaggedServiceIds('translation.loader') as $id => $attributes) {
+            $loaderRefs[$id] = new Reference($id);
             $loaders[$id][] = $attributes[0]['alias'];
             if (isset($attributes[0]['legacy-alias'])) {
                 $loaders[$id][] = $attributes[0]['legacy-alias'];
@@ -35,11 +39,15 @@ class TranslatorPass implements CompilerPassInterface
             $definition = $container->getDefinition('translation.loader');
             foreach ($loaders as $id => $formats) {
                 foreach ($formats as $format) {
-                    $definition->addMethodCall('addLoader', array($format, new Reference($id)));
+                    $definition->addMethodCall('addLoader', array($format, $loaderRefs[$id]));
                 }
             }
         }
 
-        $container->findDefinition('translator.default')->replaceArgument(2, $loaders);
+        $container
+            ->findDefinition('translator.default')
+            ->replaceArgument(0, (new Definition(ServiceLocator::class, array($loaderRefs)))->addTag('container.service_locator'))
+            ->replaceArgument(3, $loaders)
+        ;
     }
 }